\name{Pinkplots}
\alias{Pinkplots}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Pinkplots
}
\description{
Make regression and column plots that are pink!
}
\usage{
pink_col(data, x, y)
pink_regression(data, x,y, title = NULL, se = TRUE)
}
%- maybe also 'usage' for other objects documented here.

\author{
Radha Kautavarapu
%%  ~~who you are~~
}
